add file( leap year)
Write a program to check leap year using if else. How to check whether a given year is a leap year or not. [Hint:Take an input of any number. Store it in some variable say year. If a year is exactly divisible by 4 and not divisible by 100, then it is a leap year. Or if a year is exactly divisible by 400 then it is a leap year.]